+++ blomsterLandet/src/App.js
@@ -8,7 +8,7 @@ import * as firebaseHandler from './firebaseHandler';
 import NotificationScreen from './screens/NotificationScreen';
 import MyPlantScreen from './screens/MyPlantScreen';
 import ProfileScreen from './screens/ProfileScreen';
-import ShopScreen from './screens/ShopScreen';
+import ShopScreen from './screens/shop/ShopScreen';
 import DevScreen from './screens/DevScreen';
 
 export const TabNavigator = createBottomTabNavigator({
@@ -18,37 +18,37 @@ export const TabNavigator = createBottomTabNavigator({
     Profile: ProfileScreen,
     Dev: DevScreen
 },
-// Adding icons to the navigation bar.
-// Might have to change the cases if languaged in the app is changed to swedish as
-// they are based on the labels on the navigation tabs. 
-{
-    defaultNavigationOptions: ({ navigation }) => ({
-        tabBarIcon: ({ focused }) => {
-            const { routeName } = navigation.state;
-            let iconName;
-            switch (routeName) {
-                case 'Shop':
-                    iconName = 'local-grocery-store';
-                    break; 
-                case 'Plants':
-                    iconName = 'local-florist';
-                    break;
-                case 'Notifications':
-                    iconName = 'notifications';
-                    break;
-                case 'Profile':
-                    iconName = 'person';
-                    break;
-                case 'Dev':
-                    iconName = 'developer-mode';
-                    break;
-                default:   
-                    break;
+    // Adding icons to the navigation bar.
+    // Might have to change the cases if languaged in the app is changed to swedish as
+    // they are based on the labels on the navigation tabs. 
+    {
+        defaultNavigationOptions: ({ navigation }) => ({
+            tabBarIcon: ({ focused }) => {
+                const { routeName } = navigation.state;
+                let iconName;
+                switch (routeName) {
+                    case 'Shop':
+                        iconName = 'local-grocery-store';
+                        break;
+                    case 'Plants':
+                        iconName = 'local-florist';
+                        break;
+                    case 'Notifications':
+                        iconName = 'notifications';
+                        break;
+                    case 'Profile':
+                        iconName = 'person';
+                        break;
+                    case 'Dev':
+                        iconName = 'developer-mode';
+                        break;
+                    default:
+                        break;
+                }
+                return <Icon name={iconName} fontSize={30} flex={1} size={25} color={focused ? '#adc22d' : '#000'} />;
             }
-            return <Icon name={iconName} fontSize={30} flex={1} size={25} color={focused ? '#adc22d' : '#000'} />;
-        }
-    })
-}
+        })
+    }
 );
 
 const TabNavigatorContainer = createAppContainer(TabNavigator);
@@ -61,7 +61,7 @@ class App extends React.Component {
     }
     render() {
         return (
-            <ImageBackground 
+            <ImageBackground
                 // eslint-disable-next-line global-require
                 source={require('./resources/images/page-content-bg.png')}
                 style={{ width: '100%', height: '100%' }}
