+++ blomsterLandet/src/screens/shop/ShopItemPreviewScreen.js
@@ -1,14 +1,15 @@
 import React from 'react';
 import { Text, View } from 'react-native';
 
-class PlantDetailView extends React.Component {
+class ShopItemPreviewScreen extends React.Component {
     render() {
         return (
             <View style={{ flex: 1, alignItems: "center", justifyContent: "center" }}>
                 <Text>Details Screen</Text>
             </View>
+
         );
     }
 }
 
-export default PlantDetailView;
+export default ShopItemPreviewScreen;
